% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceiling_to_base.R
\name{ceiling_to_base}
\alias{ceiling_to_base}
\title{Ceiling to base}
\usage{
ceiling_to_base(x, base)
}
\arguments{
\item{x, }{number to be rounded using ceiling}

\item{base, }{number to use for rounding, e.g. to nearest 5}
}
\value{
numeric
}
\description{
Calculate ceiling for certain base level, e.g. 72 to base 5 returns 75. 
Useful for creating sensible upper limit on graph axes.
}
\examples{
ceiling_to_base(72, 5)

}
