% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{centroids002}
\alias{centroids002}
\title{Centroid coordinates of 2-km squares (tetrads)}
\format{
A data.frame with 84969 rows and 3 columns
\describe{
  \item{character}{tetrad_id (e.g. TL88A)}
  \item{numeric}{easting on British National Grid (metres)}
  \item{numeric}{northing on British National Grid (metres)}
}
}
\description{
A list of coordinates on British National Grid of the centroid of all British and Irish tetrads. Channel Islands squares are included 
but note the location has been shifted as used in Bird Atlas maps.
}
\keyword{data}
