% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{centroids010}
\alias{centroids010}
\title{Centroid coordinates of 10-km squares (hectads)}
\format{
A data.frame with 5353 rows and 3 columns
\describe{
  \item{character}{tenkm (e.g. TL88)}
  \item{numeric}{easting on British National Grid (metres)}
  \item{numeric}{northing on British National Grid (metres)}
}
}
\description{
A list of coordinates on British National Grid of the centroid of all British and Irish 10-km squares. Channel Islands squares are included 
but note the location has been shifted as used in Bird Atlas maps.
}
\keyword{data}
