% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{centroids020}
\alias{centroids020}
\title{Centroid coordinates of 20-km squares}
\format{
A data.frame with 1386 rows and 3 columns
\describe{
  \item{character}{segref (e.g. TL_A) where notation is as described in Bird Atlas p35}
  \item{numeric}{easting on British National Grid (metres)}
  \item{numeric}{northing on British National Grid (metres)}
}
}
\description{
A list of coordinates on British National Grid of the centroid of all British and Irish 20-km squares. Channel Islands squares are included 
but note the location has been shifted as used in Bird Atlas maps.
}
\keyword{data}
