% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{centroids100}
\alias{centroids100}
\title{Centroid coordinates of 100-km squares}
\format{A data.frame with 74 rows and 3 columns
\describe{
  \item{character}{hundref (e.g. TL) where notation is as described in Bird Atlas p35}
  \item{numeric}{easting on British National Grid (metres)}
  \item{numeric}{northing on British National Grid (metres)}
}}
\description{
A list of coordinates on British National Grid of the centroid of all British and Irish 100-km squares. Channel Islands squares are included 
but note the location has been shifted as used in Bird Atlas maps.
}
\keyword{data}
