% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_archive_path.R
\name{change_archive_path}
\alias{change_archive_path}
\title{Change path for archived data files}
\usage{
change_archive_path(newpath)
}
\arguments{
\item{newpath}{}
}
\description{
On load the package detects the operating environment and sets the path root
for the Unix archive space (for windows this is "//btodomain/FILES/UNIXArchive/").
Use this function to reset the path for instances where you are using data from
a different location. Note that you will still need data in the same folder
structure below this path level (e.g. ../birdatlas2007-11/data/processesd..).
}
\examples{
change_archive_path('C:/local_data_copies/')
}
