% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_gridref.R
\name{coords_to_gridref}
\alias{coords_to_gridref}
\title{Convert a set of eastings/northings to grid references}
\usage{
coords_to_gridref(df, res, outname = "gref")
}
\arguments{
\item{df}{name of input dataframe}

\item{res}{resolution of desired grid reference (100, 50, 20, 10, 2, 1km)}

\item{outname}{optional desired name of output grid reference variable (e.g. 'tenkm')}
}
\value{
The same dataframe with additional column for the requested grid reference, either
named 'gref' or optionally renamed
}
\description{
Takes a dataframe of British National grid eastings and northings and converts to standard grid
reference of desired resolution. Note this only works for locations in Great Britain.
}
\examples{
temp<-data.frame(cbc_code='R.',count=10,easting=552713.37,northing=292678.498)
temp2<-coords_to_gridref(temp,2,'tetrad_id')
print(temp2)

}
