% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid_for_object.R
\name{create_grid_for_object}
\alias{create_grid_for_object}
\title{Create grid of squares for a spatial object}
\usage{
create_grid_for_object(
  sp_object,
  grid_resolution,
  clip2object = TRUE,
  region = NULL
)
}
\arguments{
\item{sp_object}{= the polygon for which the grid is needed}

\item{grid_resolution}{= the resolution in metres of the desired grid}

\item{clip2object}{= bool, should squares with no overlap be removed?}

\item{region}{= One of GB (=Great Britain), I (=Ireland) or CH (=Channel Islands)}
}
\value{
An sf object
}
\description{
Create a grid of a given resolution for a polygon.
}
\details{
Rather than store shapefiles of grids of different sizes and projections, 
use this function to create a grid of a required resolution spanning a reigon of 
interest. Optionally limit the grid to cells overlapping sp_object.
}
