% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering_functions.R
\name{delete_1990_extralimitals}
\alias{delete_1990_extralimitals}
\title{Remove suspect breeding records from 1988-91 distribution data}
\usage{
delete_1990_extralimitals(df)
}
\arguments{
\item{df}{name of input dataframe}
}
\description{
Differences in breeding evidence criteria mean that some apparent "breeding" records in the
1988-91 atlas most likely represent migrants or non-breeders. These complicate range change
comparisons and in some cases it may be helpful to remove them. They are difficult to identify
but this function removes some obvious ones (e.g. extralimital passage waders). This function
requires species to be identified by numeric speccode variable and squares by character tenkm variable.
}
\examples{
distrib1990.clean<-delete_1990_extralimitals(distrib1990)

}
