% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering_functions.R
\name{exclude_channel_isles}
\alias{exclude_channel_isles}
\title{Remove data for the Channel Islands from a dataframe}
\usage{
exclude_channel_isles(df, var2use = "tenkm")
}
\arguments{
\item{df}{name of input dataframe}

\item{var2use}{name of grid reference style variable on which to remove records}
}
\value{
A dataframe in same format as the input dataset.
}
\description{
Remove all records for the Channel Isles from a dataframe
}
\examples{
df<-exclude_channel_isles(distrib2010,'tenkm')

}
