% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering_functions.R
\name{exclude_hybrids}
\alias{exclude_hybrids}
\title{Remove subspecies, hybrids and domestic forms from a dataframe}
\usage{
exclude_hybrids(df)
}
\arguments{
\item{df}{name of input dataframe}
}
\value{
A dataframe in same format as the input dataset.
}
\description{
Remove subspecies, hybrids and domestic forms from a dataframe. As some of these do not have a
cbc_code this can only be done using numeric speccode variable.
}
\examples{
distrib2010.truespecsonly <- exclude_hybrids(distrib2010)

}
