% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering_functions.R
\name{exclude_ireland}
\alias{exclude_ireland}
\title{Remove data for Ireland from a dataframe}
\usage{
exclude_ireland(df, var2use = "tenkm")
}
\arguments{
\item{df}{name of input dataframe}

\item{var2use}{name of grid reference style variable on which to remove records}
}
\value{
A dataframe in same format as the input dataset.
}
\description{
Remove all records for Ireland from a dataframe
}
\examples{
df<-exclude_ireland(distrib2010,'tenkm')

}
