% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering_functions.R
\name{exclude_seabirds}
\alias{exclude_seabirds}
\title{Remove colonial seabirds from a dataframe}
\usage{
exclude_seabirds(df, var2use)
}
\arguments{
\item{df}{name of input dataframe}

\item{var2use}{whether to use speccode or cbc_code to filter species}
}
\value{
A dataframe in same format as the input dataset.
}
\description{
Remove colonial seabirds from a dataframe using either speccode or cbc_code variable.
}
\examples{
distrib2010.noseabirds<-exclude_seabirds(distrib2010,'speccode')

}
