% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_list.R
\name{expand_list}
\alias{expand_list}
\title{Format a list of words with commas and 'and'}
\usage{
expand_list(inval)
}
\arguments{
\item{inval}{a vector of strings to be expanded}
}
\description{
Take a list and pad it out with commas and 'and'.
}
\details{
For use in RMarkdown documents to convert lists of species, 
for example, into grammatically correct sentences.
}
\examples{
x <- c('Robin', 'Dunnock', 'Wren')
expand_list(inval = x)

}
