% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_info.R
\name{get_species_info}
\alias{get_species_info}
\title{Get species names and codes for a taxonomic entity}
\usage{
get_species_info(field, value)
}
\arguments{
\item{field}{str = which field to search on, one of: id, name, sciname, 2ltr, 5ltr or euring}

\item{value}{str or numeric = the value to search with, e.g. 10}
}
\value{
A one-row dataframe.
}
\description{
Given a species name, scientific name, or species code, return all other names and codes for this entity.
}
\examples{
get_species_info('id', 18)
get_species_info('name', 'Blue Tit')
get_species_info('sciname', 'Emberiza calandra')
get_species_info('2ltr', 'W.')
get_species_info('5ltr', 'JAY..')
get_species_info('euring', 11870)

}
