% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{global_species_lookup}
\alias{global_species_lookup}
\title{Species dictionary lookup}
\format{A data.frame with 34051 rows and eight columns
\describe{
  \item{numeric}{Numeric taxon key code}
  \item{character}{Species' scientific name}
  \item{character}{Species' common name}
  \item{numeric}{Taxonomic rank identification (numeric) of entity}
  \item{character}{Taxonomic rank of entity}
  \item{numeric}{Sort order under IOC}
  \item{character}{Two-letter code if exists}
}}
\description{
Common names, scientific names, two-letter codes, sort order and taxonomic rank for birds, mammals and various inverts
}
\references{
IOC
}
\keyword{data}
