% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref_to_coordinates.R
\name{gridref_to_coordinates}
\alias{gridref_to_coordinates}
\title{Converts a column of British National grid references to columns of eastings and northings}
\usage{
gridref_to_coordinates(df, invar_gridref)
}
\arguments{
\item{df}{name of input dataframe}

\item{invar_gridref}{quoted string name of the variable containing the grid reference}
}
\value{
The same dataframe with additional easting and northing columns.
}
\description{
Takes a dataframe of British National grid reference and produce new columns of eastings and 
northings. Only works for Britain.
}
\examples{
temp1 <- data.frame(other1 = 'A', other2 = 'B', 
square = c('TL1', 'TL12', 'TL12A', 'TL1234', 'TL123456', 'TL1234567', 'GB1234', 'TL12O'), 
stringsAsFactors = FALSE)
temp1a<-gridref_to_coordinates(temp1,'square')

}
