% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{landarea010}
\alias{landarea010}
\title{Land area of each 10-km squares}
\format{
A data.frame with 3894 rows and 2 columns
\describe{
  \item{character}{tenkm (e.g. TL88)}
  \item{numeric}{area (square km)}
}
}
\description{
The area (in sqr km) of each 10-km square in Britain, Ireland and the Channel Islands, based on low tide line.
}
\keyword{data}
