% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering_functions.R
\name{limit2chequerboard_britain}
\alias{limit2chequerboard_britain}
\title{Limit a dataframe to the chequerboard in Britain}
\usage{
limit2chequerboard_britain(df)
}
\arguments{
\item{df}{name of input dataframe}
}
\value{
A dataframe in same format as the input dataset.
}
\description{
For some analyses it may be necessary to impose the chequerboard approach onto Britain (even though
all 10-km squares were actually targetted there).
}
\examples{
df<-limit2chequerboard_britain(irish.ttvs)

}
