% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering_functions.R
\name{limit2chequerboard_ireland}
\alias{limit2chequerboard_ireland}
\title{Limit a dataframe to the chequerboard in Ireland}
\usage{
limit2chequerboard_ireland(df)
}
\arguments{
\item{df}{name of input dataframe}
}
\value{
A dataframe in same format as the input dataset.
}
\description{
For some Atlas analyses it may be necessary to remove the 10-km squares in 
Ireland that were not on the chequerboard and hence did not achieve the 
minimum TTV coverage. The 10-km squares to remove are all those where the
easting and northing sum to an odd  number.
}
\examples{
df<-limit2chequerboard_ireland(irish.ttvs)

}
