% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_abundance_20km_2010.R
\name{load_abundance_20km_2010}
\alias{load_abundance_20km_2010}
\title{Load 20-km relative abundance data for Bird Atlas 2007-11}
\usage{
load_abundance_20km_2010(season = "BW")
}
\arguments{
\item{season}{Whether to return data for winter (W), breeding season (B) or both (BW)}
}
\value{
A dataframe with the following columns:
\item{segref}{20-km grid reference}
\item{season}{Breeding (B) or Winter (W)}
\item{speccode}{numeric species code}
\item{ntets}{number of tetrads surveyed in 20-km}
\item{freqindex}{proportion of tetrads in 20-km occupied}
\item{meancount}{mean of counts (averaged across hours, then across visits)}
}
\description{
Summarised relative abundance data from the Bird Atlas 2007-11 for all species.
Data are all at 20-km resolution and as such give confidential locations
for some rare breeding species. Note that breeding season abundance has been
removed for 10-km squares where there was no breeding evidence (e.g. for migrants).
}
\examples{
df<-load_abundance_20km_2010(season='B')

}
