% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_abundance_change_20km.R
\name{load_abundance_change_20km}
\alias{load_abundance_change_20km}
\title{Load 20-km relative abundance change data (1988-91 to 2008-11)}
\usage{
load_abundance_change_20km()
}
\value{
A dataframe with the following columns:
\item{segref}{20-km grid reference}
\item{cbc_code}{2-letter species code}
\item{index90}{proportion of tetrads occupied in 1988-91}
\item{index10}{proportion of tetrads occupied in 2008-11}
\item{diff}{arithmetic difference in proportion of tetrads occupuied between atlases}
\item{easting}{British National grid easting}
\item{northing}{British National grid northing}
}
\description{
A measure of relative abundance change based on the difference between "frequency indices"
from the 1988-91 and 2008-11 atlas datasets. Calculated on a 20-km grid to allow the same
resolution of data presentation for Britain and Ireland (where there was a lower intensity
of sampling in 2008-11).
}
\examples{
df<-load_abundance_change_20km()

}
