% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_area_10km.R
\name{load_area_10km}
\alias{load_area_10km}
\title{Load land area of 10-km square}
\usage{
load_area_10km(type = "BIlow")
}
\arguments{
\item{type}{Which version of areas to load: Britain & Ireland low tide (default) or GB high tide}
}
\value{
A dataframe with the following columns:
\item{tenkm}{10-km square}
\item{area}{area in square km}
}
\description{
Land area of the 10-km square. Two versions are available. One (type=BIlow) is based on low tide
mark and is available for Britain and Ireland. The other (type=GBhigh) is based on high tide mark
and is only available for GB.
}
\examples{
df<-load_area_10km(type=BIlow)

}
