% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_coords.R
\name{load_coords}
\alias{load_coords}
\title{Load centroid coordinates for squares of various sizes}
\usage{
load_coords(res)
}
\value{
A dataframe with the following columns:
\item{tenkm}{10-km square, or tetrad_id, segref, quadref, hundref depending on chosen grid resolution}
\item{easting}{British National grid easting of square centroid}
\item{northing}{British National grid northing of square centroid}
}
\description{
Chose to load coordinates for tetrads, 10-km, 20-km, 50-km or 100-km squares. Note that coordinates
for the Channel Isles have been moved to where they were mapped in the Atlas book.
}
\examples{
df<-load_coords(res=10)

}
