% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_distribution_1970.R
\name{load_distribution_1970}
\alias{load_distribution_1970}
\title{Load 10-km distribution data for the 1968-72 Breeding Atlas}
\usage{
load_distribution_1970()
}
\value{
A dataframe with the following columns:
\item{speccode}{numeric species code}
\item{cbc_code}{two-letter species code (for those species that have one, otherwise NA)}
\item{tenkm}{10-km grid reference of record}
\item{tenkm_published}{10-km grid reference as used in the book which may have been moved for rare breeding birds}
\item{cat}{breeding status (1 = possible breeding, 2 = probable breeding, 3 = confirmed breeding)}
}
\description{
Summarised distribution data from the 1968-72 breeding atlas for all species.
Data are all at 10-km resolution and as such give confidential locations
for some rare breeding species; for these species there may be a different square
in the "published" column. These data can be used to
replicate distribution maps or for analyses. Note some information is NA, e.g.  where we
records were apparently not published.
}
\examples{
df<-load_distribution_1970()

}
