% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_distribution_1990.R
\name{load_distribution_1990}
\alias{load_distribution_1990}
\title{Load 10-km distribution data for the 1988-91 Breeding Atlas}
\usage{
load_distribution_1990()
}
\value{
A dataframe with the following columns:
\item{speccode}{numeric species code}
\item{cbc_code}{two-letter species code (for those species that have one, otherwise NA)}
\item{tenkm}{10-km grid reference of record}
\item{tenkm_published}{10-km grid reference as used in the book which may have been moved for rare breeding birds}
\item{cat}{breeding status (S = seen, B = Breeding)}
\item{cat_published}{breeding status as used in the book, which may have been downgraded or removed}
}
\description{
Summarised distribution data from the 1988-91 breeding atlas for all species.
Data are all at 10-km resolution and as such give confidential locations
for some rare breeding species; for these species there may be a different square or
breeding evidence code in the "published" columns. These data can be used to
replicate distribution maps or for analyses. Note some information is NA, e.g.  where we
could not deduce true locations of moved records.
}
\examples{
df<-load_distribution_1990()

}
