% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_non_native_definitions.R
\name{load_non_native_definitions}
\alias{load_non_native_definitions}
\title{Load lists of species considered to be established vs exotic non-natives}
\usage{
load_non_native_definitions()
}
\value{
A dataframe with the following columns:
\item{speccode}{numeric species code}
\item{non_native_b}{species definition for breeding data = C=established, E=exotic}
\item{non_native_w}{species definition for winter data = C=established, E=exotic}
}
\description{
For some analyses it may be appropriate to distinguish between native and non-native species,
and to potentially separate the latter into those with established populations (Category C)
and those yet to establish (Category E or exotic). Some species can be safely considered
to be non-natives in the breedng season (e.g. Red-breasted Goose) whereas in winter true
wild vagrants can occur. Hence a different classification is provided for winter and breeding.
}
\examples{
df<-load_non_native_definitions()

}
