% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_ttvs_1990.R
\name{load_raw_ttvs_1990}
\alias{load_raw_ttvs_1990}
\title{Load 1988-91 breeding atlas Timed Tetrad Visit data in unsummarised format}
\usage{
load_raw_ttvs_1990()
}
\value{
A dataframe with the following columns:
\item{tenkm}{10-km square surveyed}
\item{year}{year square surveyed}
\item{tetrad_num}{no. tetrads visited}
\item{tetrad_1v}{no. tetrads visited only once}
\item{A_visits}{number of visits to tetrad A}
\item{B_visits}{number of visits to tetrad B}
\item{...}{23 further variables for number of visits to remaining 23 tetrads}
\item{cbc_code}{2-letter species code}
\item{A}{data for species in tetrad A; see Details}
\item{B}{data for species in tetrad B; see Details}
\item{...}{23 further variables for species data in remaining 23 tetrads}
\item{total_tetrad_p}{num of tetrads a species was recorded in}
\item{total_count}{total count across all tetrads visited}
\item{cat}{breeding status of species based on these TTVs (S=seen, B=breeding); see Details}
}
\description{
Loads raw TTV presence and count data for all TTVs during the 1988-91 breeding atlas. Note that
some tetrads were visited more than twice so these data cannot be used for fixed effort calculation
and instead the filtered data should be used (see load_safe_ttvs_1990). Note also that unlike
Bird Atlas 2007-11, not all species were counted - many were simply listed as present during TTVs,
hence the odd structure of 1s, 0s and counts which must be treated carefully so as not to incorrectly
infer presence or absence.
}
\details{
IMPORTANT: these data are essentially a direct transcription of the paper forms, so the species
information in columns A to Z needs some careful interpretation. A zero in these columns denotes
a tick on the form, which meant a species was present. So a zero is NOT an absence. Any other number
in these columns should be for a count species (e.g. rare passerine, waterbird, colonial species) and
indicates the maximum count for the species across TTVs.

Note also that the breeding evidence in the cat variable is based on observations made during the TTVs
only. Subsequent casual records will likely have increased the breeding evidence. This column isn't
really useful for the majority of analyses.
}
\examples{
df<-load_raw_ttvs_1990()

}
