% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_specnames.R
\name{load_specnames}
\alias{load_specnames}
\title{Load species codes and names as used in Bird Atlas 2007-11}
\usage{
load_specnames()
}
\value{
A dataframe with the following columns:
\item{speccode}{numeric species code}
\item{cbc_code}{two-letter species code (for those species that have one, otherwise NA)}
\item{engname}{species common name}
\item{sciname}{species scientific name}
\item{bou_order}{species order at time of Atlas}
}
\description{
A lookup file of species codes and English and scientific names as used in Bird Atlas 2007-11. Note that
scientific names and taxonomic order has changed since.
}
\examples{
df<-load_specnames()

}
