% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ttv_details_2010.R
\name{load_ttv_details_2010}
\alias{load_ttv_details_2010}
\title{Load visit information for Bird Atlas 2007-11 TTVs}
\usage{
load_ttv_details_2010(trim2010 = TRUE)
}
\arguments{
\item{trim2010}{Whether to remove data for local atlas TTVs completed outside the main Bird Atlas period (defaults to TRUE)}
}
\value{
A dataframe with the following columns:
\item{btoreg}{4-letter code for Atlas region}
\item{user_id}{Oracle observer username}
\item{tenkm}{10-km square}
\item{tetlet}{tetrad (A:Z minus O)}
\item{datestring}{visit date as a string, e.g. 19122007}
\item{source}{W=web submission; P=Paper submission}
\item{duration}{TTV duration: 1 or 2 hours}
\item{obsdt}{observation date as a Date object}
}
\description{
Load the observer, date and duration information for all Timed Tetrad Visits
from Bird Atlas 2007-11.
}
\examples{
df<-load_ttv_details_2010()

}
