% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ttv_matrix_firsthour_2010.R
\name{load_ttv_matrix_firsthour_2010}
\alias{load_ttv_matrix_firsthour_2010}
\title{Load Bird Atlas 2007-11 species-by-visit matrix of first-hour TTV counts}
\usage{
load_ttv_matrix_firsthour_2010()
}
\value{
A dataframe with the following columns:
\item{user_id}{Oracle observer username}
\item{tenkm}{10-km square}
\item{tetlet}{tetrad letter (A:Z, minus O)}
\item{obsyear}{year of visit}
\item{obsmonth}{month of visit}
\item{obsday}{day of visit}
\item{season}{Breeding (B) or Winter (W) season}
\item{X_###}{334 columns, one per species, containing 1st hour count. Columns names are numeric species codes with prefix X_ (e.g. X_105)}
}
\description{
A large matrix with TTV visits as rows, species as columns, giving the first-hour count.
Species not detected on visits are populated with zero counts. Note that counts outside
the range have not been manipulated.
}
\examples{
df<-load_ttv_matrix_firsthour_2010()

}
