% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2text.R
\name{num2text}
\alias{num2text}
\title{Handle numbers, text and singular or plural verbs}
\usage{
num2text(x, unit = NULL, st = FALSE)
}
\arguments{
\item{x}{= the number to be formated}

\item{unit}{= text string of the singular form of unit, e.g. '10-km square'}

\item{st}{= whether the string will be used at the start of a sentence. Determines conversion of number and capitalisation}
}
\value{
= a string containing the count as a numeral or text and a verb or unit, pluralised as required.
}
\description{
Convert numbers to text and pluralise units on demand.
}
\examples{
num2text(x = 7, unit = 'species', st = FALSE) 
num2text(x = 81, unit = 'tetrad', st = TRUE) 
num2text(x = 1547, unit = '1-km square', st = FALSE) 

}
