% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_geometry.R
\name{rename_geometry}
\alias{rename_geometry}
\title{Rename the geometry column of an sf object}
\usage{
rename_geometry(sf_object, name = "geometry")
}
\arguments{
\item{sf_object}{= the sf object to be corrected}

\item{name}{= the name to be used, defaulting to 'geometry'}
}
\description{
Rename geometry column of an sf object
}
\details{
Sometimes for unknown reason the geometry column of an sf object can be 
named 'x'instead of the expected 'geometry'. This function corrects this.
}
