% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_1km_to_10km.R
\name{rescale_1km_to_10km}
\alias{rescale_1km_to_10km}
\title{Convert a 1-km grid reference to a 10-km grid reference}
\usage{
rescale_1km_to_10km(df, invar)
}
\arguments{
\item{df}{name of input dataframe}

\item{invar}{name of the variable containing the input grid reference}
}
\value{
The same dataframe with additional tenkm column
}
\description{
Takes a dataframe of British National grid 1-km references and converts to 10-km references.
Works for Britain, Ireland and Channel Isles grids
}
\examples{
temp1<-data.frame(cbc_code='R.',count=10,onekm='TL1234', stringsAsFactors = FALSE)
rescale_1km_to_10km(temp1,'onekm')

}
