% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_1km_to_10km.R
\name{rescale_1km_to_10km_}
\alias{rescale_1km_to_10km_}
\title{Convert a 1-km grid reference to a 10-km grid reference}
\usage{
rescale_1km_to_10km_(x)
}
\arguments{
\item{x}{character vector containing the input grid reference}
}
\value{
A character vector with tenkm references
}
\description{
Takes a vector of British National grid 1-km references and converts to 10-km references.
Works for Britain, Ireland and Channel Isles grids
}
\examples{
temp1<-data.frame(cbc_code='R.',count=10,onekm='TL1234', stringsAsFactors = FALSE)
rescale_1km_to_10km_(temp1$onekm)

}
