% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_1km_to_2km.R
\name{rescale_1km_to_2km}
\alias{rescale_1km_to_2km}
\title{Convert a 1-km grid reference to a 2-km (tetrad) grid reference}
\usage{
rescale_1km_to_2km(df, invar)
}
\arguments{
\item{df}{name of input dataframe}

\item{invar}{name of the variable containing the input grid reference}
}
\value{
The same dataframe with additional tetrad_id column.
}
\description{
Takes a dataframe of British National grid 1-km references and converts to 2-km references
(i.e. tetrads). Works for Britain, Ireland and Channel Isles grids
}
\examples{
temp<-data.frame(cbc_code='R.',count=10,onekm='TL1234')
temp2<-rescale_1km_to_2km(temp,'onekm')
print(temp2)
cbc_code count  onekm tetrad_id
      R.    10 TL1234    TL13H

}
