% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_1km_to_2km.R
\name{rescale_1km_to_2km_}
\alias{rescale_1km_to_2km_}
\title{Convert a 1-km grid reference to a 2-km (tetrad) grid reference}
\usage{
rescale_1km_to_2km_(x)
}
\arguments{
\item{x}{string input variable}
}
\value{
A vector with tetrad references.
}
\description{
Takes a vector of British National grid 1-km references and converts to 2-km references
(i.e. tetrads). Works for Britain, Ireland and Channel Isles grids
}
\examples{
temp1<-data.frame(cbc_code='R.',count=10,onekm='TL1234', stringsAsFactors = FALSE)
rescale_1km_to_2km_(temp1$onekm)

}
