% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{speccode_mapping}
\alias{speccode_mapping}
\title{Species code mapping}
\format{
A data.frame with 462 rows and two columns
\describe{
  \item{numeric}{old_species_code = Numeric speceies code as used in older archived files}
  \item{numeric}{new_species_code = Numeric speceies code as used currently in Oracle etc}
}
}
\description{
Following the move to use IOC world bird names, some species codes had to be updated in the Oracle system. 
However, older archived files (e.g. those for BirdAtlas 2007-11 and earlier ) still have the old species code 
encoding. For example Cattle Egret will be represented in old files as speccode = 35 whereas to match with 
new species dictionary information this needs to be updated to 52121. This file provides the necessary lookup
and can be used in conjunction with the BTOTools update_speccode() function. Based on  IOC version 11.2, 
last updated August 2021
}
\keyword{data}
