% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{squares_01km_northernireland}
\alias{squares_01km_northernireland}
\title{1-km squares in Northern Ireland}
\format{
A data.frame with 16102 rows and two columns
\describe{
  \item{character}{onekm = 1-km grid reference}
  \item{character}{country = country name}
}
}
\description{
List of 1-km squares with land in Northern Ireland (see also lists for other countries. 
A square is listed for a country if it contains any land. Therefore, individual 
squares can be present in more than one list if they fall on a country border. 
Lists are based on low tide shapefiles so include some sand flats.
}
\keyword{data}
