% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{squares_10km_dominant_country}
\alias{squares_10km_dominant_country}
\title{10-km squares assigned to dominant country}
\format{
A data.frame with 4154 rows and two columns
\describe{
  \item{character}{country = country name}
  \item{character}{tenkm = 10-km grid reference}
}
}
\description{
List of 10-km squares with land in Britain and Ireland where each square is 
assigned to a single country based on dominant land area in the square. 
Therefore, individual squares should be present in only one country.
Lists are based on low tide shapefiles so include some sand flats. Includes 
Isle of Man (assigned England) but not Channel Islands.
}
\keyword{data}
