% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_scientific_names.R
\name{update_scientific_names}
\alias{update_scientific_names}
\title{Update species' scientific names}
\usage{
update_scientific_names(df, var)
}
\arguments{
\item{df}{character The name of the dataset to be scanned for name updates}

\item{var}{character A quoted string. The name of the column in df to be scanned for name updates}
}
\value{
A dataframe
}
\description{
The function does not perform taxonomic changes such as splits and lumps. All it does is change 
names where there have been spelling changes (e.g. montanus to montana), or where genus names 
have changed (e.g. Parus to Periparus). It uses a lookup table of species synonyms which is 
available within the package and will be updated periodically.
}
\examples{
old_records <- data.frame(sciname = c('Apus apus', 'Parus ater'), engname = c('Swift', 'Coal Tit'), stringsAsFactors = FALSE)
old_records_updated <- update_scientific_names(df = old_records, var = 'sciname')

}
\seealso{
\code{\link{sciname_synonyms}}
}
