% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_speccodes.R
\name{update_speccodes}
\alias{update_speccodes}
\title{Update old species' numeric codes}
\usage{
update_speccodes(df, var, report_changes = FALSE)
}
\arguments{
\item{df}{character The name of the dataset in which speccodes will be updated}

\item{var}{character A quoted string. The name of the column containing species codes to be updated}

\item{report_changes}{logical Should the list of changes be displayed on execution (default FALSE)}
}
\value{
A dataframe.
}
\description{
The function does not perform taxonomic changes such as splits and lumps. All it does is change 
the species code used in old archived files to the newer version used everywhere else. This then 
allows use of the global_species_dictionary. It uses a lookup table of species code mappings 
which is available within the package.
}
\details{
If \code{report_changes=TRUE} the list of cases changed is printed to output, showing original
name (from), destination name (to) and the number of rows affected.
}
\examples{
old_records <- data.frame(speccode = c(36, 153),
                          engname = c('Cattle Egret', 'Black-winged Stilt'),
                          stringsAsFactors = FALSE)
old_records_updated <- update_speccodes(df = old_records,
                                             var = 'speccode')

}
\seealso{
\code{\link{speccode_mapping}}
}
