% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid_for_object.R
\name{create_grid_for_object}
\alias{create_grid_for_object}
\title{Create grid of squares for a spatial object}
\usage{
create_grid_for_object(
  sp_object,
  grid_resolution,
  clip_tolerance = 0,
  region = NULL,
  outvarname
)
}
\arguments{
\item{sp_object}{= the polygon for which the grid is needed}

\item{grid_resolution}{= the resolution in metres of the desired grid}

\item{clip_tolerance}{= numeric, the minimum overlap (sq metres) between a grid cell and sp_object for the cell to be retained}

\item{region}{= One of GB (=Great Britain), I (=Ireland) or CH (=Channel Islands)}

\item{outvarname}{= string, name for the exported grid reference column (e.g. 'tetrad_id')}
}
\value{
An sf object
}
\description{
Create a grid of a given resolution for a polygon.
}
\details{
Rather than store shapefiles of grids of different sizes and projections, 
use this function to create a grid of a required resolution spanning a reigon of 
interest. Optionally limit the grid to cells overlapping sp_object.
}
