% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{global_species_lookup}
\alias{global_species_lookup}
\title{Species dictionary lookup}
\format{
A data.frame with 34290 rows and 11 columns
\describe{
  \item{numeric}{master_taxon_id = Numeric taxon key code}
  \item{character}{scientific_name = Species' scientific name}
  \item{character}{english_name = Species' common name as in common usage in UK}
  \item{numeric}{taxon_rank_id = Taxonomic rank identification (numeric) of entity}
  \item{character}{taxon_rank_name = Taxonomic rank of entity}
  \item{numeric}{sort_order = Sort order under IOC}
  \item{character}{alt_int_name = Species' common name as used internationally (e.g. with Eurasian prefix)}
  \item{character}{code2ltr = two-letter code if exists. To be used in preference to cbc_code}
  \item{character}{code5ltr = five-letter code, if exists}
  \item{numeric}{euring = Euring number if exists}
  \item{character}{taxa} = Common name of species group (e.g. Birds, Butterflies, Reptiles)
}
}
\description{
Common names, scientific names, two-letter codes, five-letter codes, sort order and taxonomic rank for birds, 
mammals and various inverts. Now with 5-letter codes for all European breeding birds. Bird names as used in 
IOC version 12.2, last updated October 2023.
}
\references{
IOC
}
\keyword{data}
