% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_10km_to_20km.R
\name{rescale_10km_to_20km}
\alias{rescale_10km_to_20km}
\title{Convert a 10-km grid reference to a 20-km grid reference}
\usage{
rescale_10km_to_20km(df, invar)
}
\arguments{
\item{df}{name of input dataframe}

\item{invar}{name of the variable containing the input grid reference}
}
\value{
The same dataframe with additional segref column.
}
\description{
Takes a dataframe of British National grid 10-km references and converts to 20-km references.
Works for Britain, Ireland and Channel Isles grids
}
\examples{
temp1<-data.frame(cbc_code = c('R.','CH'), count = c(10,2), tenkm = c('TL13', 'SK87'), stringsAsFactors = FALSE)
rescale_10km_to_20km(temp1,'tenkm')

}
