% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_10km_to_20km.R
\name{rescale_10km_to_20km}
\alias{rescale_10km_to_20km}
\title{Convert a 10-km grid reference to a 20-km grid reference}
\usage{
rescale_10km_to_20km(df, invar)
}
\arguments{
\item{df}{name of input dataframe}

\item{invar}{name of the variable containing the input grid reference}
}
\value{
The same dataframe with additional segref column.
}
\description{
Takes a dataframe of British National grid 10-km references and converts to 20-km references.
Works for Britain, Ireland and Channel Isles grids
}
\examples{
temp1<-data.frame(cbc_code = 'R.', count = 10, tenkm = 'TL13', stringsAsFactors = FALSE)
temp1a<-rescale_10km_to_20km(temp1,'tenkm')
temp2<-data.frame(cbc_code = 'R.', count = 10, tenkm = 'TL1', stringsAsFactors = FALSE)
temp2a<-rescale_10km_to_20km(temp2,'tenkm')
temp3<-data.frame(cbc_code = 'R.', count = 10, tenkm = 'TL1234', stringsAsFactors = FALSE)
temp3a<-rescale_10km_to_20km(temp3,'tenkm')

}
