% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_10km_to_50km.R
\name{rescale_10km_to_50km}
\alias{rescale_10km_to_50km}
\title{Convert a 10-km grid reference to a 50-km grid reference}
\usage{
rescale_10km_to_50km(df, invar)
}
\arguments{
\item{df}{name of input dataframe}

\item{invar}{name of the variable containing the input grid reference}
}
\value{
The same dataframe with additional quadref column.
}
\description{
Takes a dataframe of British National grid 10-km references and converts to 50-km references.
Works for Britain, Ireland and Channel Isles grids
}
\examples{
temp<-data.frame(cbc_code='R.',count=10,tenkm='TL13')
temp2<-rescale_10km_to_50km(temp,'tenkm')
print(temp2)
cbc_code count  tenkm quadref
      R.    10   TL13    TLSW

}
